Change spikesorting schema name
